New structure for folders and classes
